refactor(parts): move reactive plugin under parts
